Update
- Se separaron las rutas en blueprints individuales.
- Se centralizó la conexión a Supabase en api/utils/supabase.py.
- Se corrigieron imports para evitar errores de variables de entorno y supabase_url.
- Ajuste en url_for para usar los nombres correctos de endpoints según blueprints.
- Limpieza general de código para mejor mantenibilidad.
