+++ api/__init__.py
@@ -1,24 +1,31 @@
-import os
 from flask import Flask
+from flask_wtf.csrf import CSRFProtect
+from api.routes import all_blueprints
+
+from dotenv import load_dotenv
+import os
+from supabase import create_client
+
+load_dotenv()
+
+SUPABASE_URL = os.getenv("SUPABASE_URL")
+SUPABASE_KEY = os.getenv("SUPABASE_KEY")
+
+if not SUPABASE_URL or not SUPABASE_KEY:
+    raise ValueError("Faltan variables de entorno SUPABASE_URL o SUPABASE_KEY")
+
+supabase = create_client(SUPABASE_URL, SUPABASE_KEY)
 
 def create_app():
+    app = Flask(__name__)
+    app.secret_key = os.getenv("SECRET_KEY", "dev_key")
 
-    # Definir la ruta al directorio de plantillas y estáticos
-    current_dir = os.path.dirname(os.path.abspath(__file__))
-    template_dir = os.path.join(current_dir, 'templates')
-    static_dir = os.path.join(current_dir, 'static')
+    CSRFProtect(app)
 
+    for bp in all_blueprints:
+        app.register_blueprint(bp)
 
-    # Crear la aplicación Flask
-    app = Flask(__name__, 
-               template_folder=template_dir,
-               static_folder=static_dir)
-
-    # Configuración de clave secreta para CSRF y sesiones
-    app.config['SECRET_KEY'] = os.environ.get('SECRET_KEY', 'clave-muy-secreta-y-unica')
-
-    # Registrar blueprints
-    from .routes import main_bp
-    app.register_blueprint(main_bp)
+    # Podés pasar supabase como global o adjuntarlo al app
+    app.supabase = supabase
 
     return app
