+++ api/routes/__init__.py
@@ -1,0 +1,5 @@
+from .main import main_bp
+from .encuesta import encuesta_bp
+from .auth import auth_bp
+
+all_blueprints = [main_bp, encuesta_bp, auth_bp]
