+++ api/routes/main.py
@@ -1,0 +1,11 @@
+from flask import Blueprint, render_template
+
+main_bp = Blueprint('main', __name__)
+
+@main_bp.route('/')
+def index():
+    return render_template('index.html', nombre='Usuario')
+
+@main_bp.route('/about')
+def about():
+    return render_template('about.html', titulo='Acerca de', contenido="Esta es una demo con Flask y Supabase en Vercel.")
