+++ api/templates/index.html
@@ -14,8 +14,8 @@
         <li>Open source y creada por latinoamericanos</li>
       </ul>
       <div class="btn-group" style="padding: 2rem 0;">
-        <a href="{{ url_for('main.encuesta') }}" class="btn">Dejar opinión</a>
-        <a href="{{ url_for('main.user_creation') }}" class="btn btn-outline">Crear cuenta</a>
+        <a href="{{ url_for('encuesta.encuesta') }}" class="btn">Dejar opinión</a>
+        <a href="{{ url_for('auth.user_creation') }}" class="btn btn-outline">Crear cuenta</a>
         <a href="{{ url_for('main.about') }}" class="btn btn-secondary">Acerca de</a>
       </div>
     </section>
